refactor(audit): clear remaining structural findings (#5645)
