+++ src/core/rig/spec.rs
@@ -9,7 +9,19 @@ use std::collections::{BTreeMap, HashMap};
 use crate::core::component::ScopedExtensionConfig;
 use crate::core::extension::bench::{BenchGate, BenchGateOp};
 
+mod check;
+mod pipeline;
+mod trace;
 mod workload;
+
+pub use check::{CheckSpec, NewerThanSpec, TimeSource};
+pub use pipeline::{GitOp, PatchOp, PipelineStep, ServiceOp, SharedPathOp, StackOp, SymlinkOp};
+pub use trace::{
+    TraceDependencySpec, TraceExperimentArtifactSpec, TraceExperimentCommandSpec,
+    TraceExperimentSpec, TraceGuardrailSpec, TraceNativePublicPreviewSpec,
+    TracePreviewAssetFanoutSpec, TraceProfileSpec, TracePublicPreviewMode, TracePublicPreviewSpec,
+    TraceVariantOverlaySpec, TraceVariantSpec,
+};
 
 /// A rig: components + services + pipelines.
 ///
@@ -620,251 +632,6 @@ fn merge_defaults_map<T: Clone>(target: &mut HashMap<String, T>, defaults: &Hash
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct TraceDependencySpec {
-    pub id: String,
-    pub kind: String,
-    pub source: String,
-
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub path: Option<String>,
-
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub plugin_file: Option<String>,
-
-    #[serde(default)]
-    pub requires_built_assets: bool,
-
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub required_paths: Vec<String>,
-
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub source_url: Option<String>,
-
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub version: Option<String>,
-
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub r#ref: Option<String>,
-
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub package_marker: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct TraceVariantSpec {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub component: Option<String>,
-
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub overlay: Option<String>,
-
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub overlays: Vec<TraceVariantOverlaySpec>,
-
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub trace_guardrails: Vec<TraceGuardrailSpec>,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct TraceProfileSpec {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub component: Option<String>,
-
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub scenario: Option<String>,
-
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub rig: Option<String>,
-
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub overlays: Vec<String>,
-
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub variants: Vec<String>,
-
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub settings: BTreeMap<String, serde_json::Value>,
-
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub public_preview: Option<TracePublicPreviewSpec>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct TracePublicPreviewSpec {
-    /// Preview provider lifecycle mode. Defaults to the existing external
-    /// command/public-origin behavior.
-    #[serde(default)]
-    pub mode: TracePublicPreviewMode,
-
-    /// Local HTTP origin to expose, for example `http://127.0.0.1:8888`.
-    pub local_origin: String,
-
-    /// Optional already-known public origin. When omitted, Homeboy starts
-    /// `command` and reads the first HTTPS URL printed to stdout.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub public_origin: Option<String>,
-
-    /// Long-running shell command that starts the tunnel provider.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub command: Option<String>,
-
-    /// Fail before the trace runner starts unless the effective origin is HTTPS.
-    #[serde(default)]
-    pub require_https: bool,
-
-    /// Human-readable provider label for artifacts.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub provider: Option<String>,
-
-    /// Seconds to wait for the provider command to print a public HTTPS URL.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub startup_timeout_seconds: Option<u64>,
-
-    /// Public-origin-relative asset URLs that must load before trace collection starts.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub required_asset_paths: Vec<String>,
-
-    /// Optional concurrent static-asset fanout check for public preview tunnels.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub asset_fanout: Option<TracePreviewAssetFanoutSpec>,
-
-    /// Homeboy-native preview tunnel settings used when `mode` is `homeboy_native`.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub native: Option<TraceNativePublicPreviewSpec>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct TracePreviewAssetFanoutSpec {
-    /// Public-origin-relative asset URLs to fetch concurrently through the preview origin.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub asset_paths: Vec<String>,
-
-    /// Maximum number of concurrent fetch workers.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub concurrency: Option<usize>,
-
-    /// Number of times to request each asset path.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub repeat_count: Option<usize>,
-
-    /// Optional body substring that every successful response must contain.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub expected_body_contains: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-#[derive(Default)]
-pub enum TracePublicPreviewMode {
-    #[default]
-    External,
-    HomeboyNative,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
-#[serde(deny_unknown_fields)]
-pub struct TraceNativePublicPreviewSpec {
-    /// Deterministic public host reserved by the native preview ingress.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub public_host: Option<String>,
-
-    /// Operator-owned wildcard domain used to derive `{session}-tunnel.<domain>`.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub operator_domain: Option<String>,
-
-    /// Stable session ID used for host generation and native client metadata.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub session_id: Option<String>,
-
-    /// Native preview ingress/broker URL consumed by `homeboy tunnel preview-client start`.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub ingress_url: Option<String>,
-
-    /// Environment variable name that supplies the native preview tunnel token.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub token_env: Option<String>,
-
-    /// Optional path to a Homeboy binary that implements the preview-client command.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub client_binary: Option<String>,
-}
-
-impl TraceProfileSpec {
-    pub fn string_settings(&self) -> Vec<(String, String)> {
-        self.settings
-            .iter()
-            .filter_map(|(key, value)| {
-                value
-                    .as_str()
-                    .map(|string| (key.clone(), string.to_string()))
-            })
-            .collect()
-    }
-
-    pub fn json_settings(&self) -> Vec<(String, serde_json::Value)> {
-        self.settings
-            .iter()
-            .filter(|(_, value)| !value.is_string())
-            .map(|(key, value)| (key.clone(), value.clone()))
-            .collect()
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct TraceGuardrailSpec {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub label: Option<String>,
-
-    #[serde(flatten)]
-    pub check: CheckSpec,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct TraceVariantOverlaySpec {
-    pub component: String,
-    pub overlay: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct TraceExperimentSpec {
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub setup: Vec<TraceExperimentCommandSpec>,
-
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub teardown: Vec<TraceExperimentCommandSpec>,
-
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub settings: BTreeMap<String, serde_json::Value>,
-
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub env: BTreeMap<String, String>,
-
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub artifacts: Vec<TraceExperimentArtifactSpec>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct TraceExperimentCommandSpec {
-    pub command: String,
-
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub cwd: Option<String>,
-
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub env: BTreeMap<String, String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum TraceExperimentArtifactSpec {
-    Path(String),
-    Detailed { label: String, path: String },
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -1355,485 +1122,6 @@ pub struct SharedPathSpec {
     pub link: String,
     /// Existing path to borrow, usually the primary checkout's dependency dir.
     pub target: String,
-}
-
-/// A pipeline step. Flat enum via `kind` discriminator so specs stay readable.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(tag = "kind", rename_all = "kebab-case")]
-pub enum PipelineStep {
-    /// Start/stop/health-check a declared service.
-    Service {
-        /// Optional stable node ID for dependency-aware pipeline ordering.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        step_id: Option<String>,
-        /// Step IDs that must run before this step.
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        depends_on: Vec<String>,
-        /// Service ID (must exist in `services`).
-        id: String,
-        /// Operation: `start`, `stop`, or `health`.
-        op: ServiceOp,
-    },
-
-    /// Delegate to `homeboy build`.
-    ///
-    /// Rigs should prefer `build` over `command` for component builds so they
-    /// pick up the component's declared `scripts.build`, extension hooks, and
-    /// error-formatting surface instead of shelling out blindly. Component
-    /// path is resolved from the rig's `components` map, so the component
-    /// doesn't need to be registered in homeboy's global registry.
-    Build {
-        /// Optional stable node ID for dependency-aware pipeline ordering.
-        #[serde(default, rename = "id", skip_serializing_if = "Option::is_none")]
-        step_id: Option<String>,
-        /// Step IDs that must run before this step.
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        depends_on: Vec<String>,
-        /// Component ID — must exist in the rig's `components` map.
-        component: String,
-        /// Human-readable label shown during execution.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        label: Option<String>,
-    },
-
-    /// Delegate a component lifecycle operation to its configured extension.
-    ///
-    /// V1 intentionally exposes only operations that Homeboy core already knows
-    /// how to dispatch through extension infrastructure. Use `command` for
-    /// one-off shell escape hatches; add new extension ops only when the
-    /// extension layer owns the corresponding lifecycle contract.
-    Extension {
-        /// Optional stable node ID for dependency-aware pipeline ordering.
-        #[serde(default, rename = "id", skip_serializing_if = "Option::is_none")]
-        step_id: Option<String>,
-        /// Step IDs that must run before this step.
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        depends_on: Vec<String>,
-        /// Component ID — must exist in the rig's `components` map.
-        component: String,
-        /// Extension-owned operation. V1 supports `build`.
-        op: String,
-        /// Human-readable label shown during execution.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        label: Option<String>,
-    },
-
-    /// Delegate to `homeboy git`.
-    ///
-    /// Wraps homeboy's own git primitive with a path override so rigs can
-    /// operate on unregistered checkouts. Supports the subset of operations
-    /// rigs actually need (MVP): `status`, `pull`, `fetch`, `checkout`,
-    /// `current-branch`. More can land as follow-up.
-    Git {
-        /// Optional stable node ID for dependency-aware pipeline ordering.
-        #[serde(default, rename = "id", skip_serializing_if = "Option::is_none")]
-        step_id: Option<String>,
-        /// Step IDs that must run before this step.
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        depends_on: Vec<String>,
-        /// Component ID — must exist in the rig's `components` map.
-        component: String,
-        /// Operation name.
-        op: GitOp,
-        /// Extra git arguments, appended after the op-specific base args
-        /// (e.g. `pull` with `["origin", "trunk"]` runs `git pull origin trunk`).
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        args: Vec<String>,
-        /// Human-readable label shown during execution.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        label: Option<String>,
-    },
-
-    /// Delegate to a declared component's stack spec.
-    ///
-    /// This is intentionally explicit: rigs only rewrite combined-fixes
-    /// branches when a pipeline author opts into a `stack` step (or the user
-    /// runs `homeboy rig sync`). `rig up` never syncs stacks implicitly.
-    Stack {
-        /// Optional stable node ID for dependency-aware pipeline ordering.
-        #[serde(default, rename = "id", skip_serializing_if = "Option::is_none")]
-        step_id: Option<String>,
-        /// Step IDs that must run before this step.
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        depends_on: Vec<String>,
-        /// Component ID — must exist in the rig's `components` map and declare
-        /// a `stack` field.
-        component: String,
-        /// Stack operation.
-        op: StackOp,
-        /// Print what WOULD happen without mutating the stack spec or target
-        /// branch. Only meaningful for `op = "sync"` today.
-        #[serde(default)]
-        dry_run: bool,
-        /// Human-readable label shown during execution.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        label: Option<String>,
-    },
-
-    /// Run an arbitrary shell command — escape hatch for operations that
-    /// don't map to a homeboy primitive (waits, custom tooling, probes).
-    ///
-    /// Prefer `build` / `git` / `check` over `command` wherever they fit:
-    /// typed steps pick up homeboy's existing error mapping, extension
-    /// hooks, and registry awareness for free.
-    Command {
-        /// Optional stable node ID for dependency-aware pipeline ordering.
-        #[serde(default, rename = "id", skip_serializing_if = "Option::is_none")]
-        step_id: Option<String>,
-        /// Step IDs that must run before this step.
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        depends_on: Vec<String>,
-        /// Shell command to execute. Runs via `sh -c` (or `cmd /C` on Windows).
-        #[serde(rename = "command")]
-        cmd: String,
-        /// Working directory. Supports variable expansion.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        cwd: Option<String>,
-        /// Env vars (merged over inherited environment).
-        #[serde(default, skip_serializing_if = "HashMap::is_empty")]
-        env: HashMap<String, String>,
-        /// Human-readable label shown during execution. If absent, `cmd` is used.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        label: Option<String>,
-    },
-
-    /// Run a shell command only when a filesystem path is missing.
-    ///
-    /// This keeps common bootstrap guards declarative without teaching Homeboy
-    /// package-manager-specific install semantics.
-    CommandIfMissing {
-        /// Optional stable node ID for dependency-aware pipeline ordering.
-        #[serde(default, rename = "id", skip_serializing_if = "Option::is_none")]
-        step_id: Option<String>,
-        /// Step IDs that must run before this step.
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        depends_on: Vec<String>,
-        /// Path whose absence triggers the command. Relative paths resolve
-        /// against `cwd` when set, otherwise against the process cwd.
-        missing: String,
-        /// Shell command to execute when `missing` does not exist.
-        #[serde(rename = "command")]
-        cmd: String,
-        /// Working directory. Supports variable expansion.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        cwd: Option<String>,
-        /// Env vars (merged over inherited environment).
-        #[serde(default, skip_serializing_if = "HashMap::is_empty")]
-        env: HashMap<String, String>,
-        /// Human-readable label shown during execution. If absent, `cmd` is used.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        label: Option<String>,
-    },
-
-    /// Declarative local requirement for common rig preflight checks.
-    ///
-    /// Use this for simple filesystem/component-path/executable assertions
-    /// instead of inline shell. When `prepare_command` is set, it runs only in
-    /// listed `prepare_phases`; normal `check` remains read-only and reports
-    /// the declared remediation.
-    Requirement {
-        /// Optional stable node ID for dependency-aware pipeline ordering.
-        #[serde(default, rename = "id", skip_serializing_if = "Option::is_none")]
-        step_id: Option<String>,
-        /// Step IDs that must run before this step.
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        depends_on: Vec<String>,
-        /// Existing path. Relative paths resolve against `cwd` when set.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        path: Option<String>,
-        /// Existing regular file.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        file: Option<String>,
-        /// Existing directory.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        dir: Option<String>,
-        /// Component whose resolved path must contain `component_path_contains`.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        component: Option<String>,
-        /// Required substring in the resolved component path.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        component_path_contains: Option<String>,
-        /// Executable name or path. Bare names are resolved from PATH.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        executable: Option<String>,
-        /// Environment variable whose value overrides `executable` lookup.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        executable_env: Option<String>,
-        /// Additional environment variables to try, in order, before PATH lookup.
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        executable_env_aliases: Vec<String>,
-        /// Command to run when the filesystem requirement is missing and the
-        /// current pipeline name appears in `prepare_phases`.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        prepare_command: Option<String>,
-        /// Pipeline names allowed to run `prepare_command`.
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        prepare_phases: Vec<String>,
-        /// Working directory for `prepare_command` and relative paths.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        cwd: Option<String>,
-        /// Env vars (merged over inherited environment) for `prepare_command`.
-        #[serde(default, skip_serializing_if = "HashMap::is_empty")]
-        env: HashMap<String, String>,
-        /// Human remediation shown when the requirement is not satisfied.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        remediation: Option<String>,
-        /// Human-readable label shown during execution.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        label: Option<String>,
-    },
-
-    /// Ensure a declared symlink exists (or verify it in `check` pipelines).
-    Symlink {
-        /// Optional stable node ID for dependency-aware pipeline ordering.
-        #[serde(default, rename = "id", skip_serializing_if = "Option::is_none")]
-        step_id: Option<String>,
-        /// Step IDs that must run before this step.
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        depends_on: Vec<String>,
-        /// Operation: `ensure` or `verify`.
-        op: SymlinkOp,
-    },
-
-    /// Ensure, verify, or clean up declared shared dependency paths.
-    SharedPath {
-        /// Optional stable node ID for dependency-aware pipeline ordering.
-        #[serde(default, rename = "id", skip_serializing_if = "Option::is_none")]
-        step_id: Option<String>,
-        /// Step IDs that must run before this step.
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        depends_on: Vec<String>,
-        /// Operation: `ensure`, `verify`, or `cleanup`.
-        op: SharedPathOp,
-    },
-
-    /// Apply (or verify) an idempotent local-only patch to a file in a
-    /// component checkout. Use case: upstream-source patches that can't be
-    /// committed to the consumer branch because rebases would carry them
-    /// to every fresh checkout. Examples: TSRMLS_CC fallback macros on
-    /// upstream Playground C sources, build-time tweaks that aren't
-    /// upstream yet.
-    ///
-    /// Idempotency is marker-based: if `marker` is already present in
-    /// the file, the step is a no-op. If the optional `after` anchor is
-    /// set and absent from the file, the step errors instead of guessing
-    /// where to insert (file structure changed → fail loudly).
-    ///
-    /// In `verify` mode the step only confirms the marker is present
-    /// without modifying — mirrors how a `check` pipeline would `grep`
-    /// for it. Use this in `check` pipelines so a stale or unpatched
-    /// checkout is reported as a failure, not silently re-patched.
-    Patch {
-        /// Optional stable node ID for dependency-aware pipeline ordering.
-        #[serde(default, rename = "id", skip_serializing_if = "Option::is_none")]
-        step_id: Option<String>,
-        /// Step IDs that must run before this step.
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        depends_on: Vec<String>,
-        /// Component ID — must exist in the rig's `components` map.
-        component: String,
-        /// File to patch, relative to the component's path. Tilde +
-        /// `${components.X.path}` / `${env.VAR}` expansion applies.
-        file: String,
-        /// Substring that uniquely identifies this patch in the file.
-        /// Used as the idempotency key — present means "already patched."
-        marker: String,
-        /// Optional anchor: substring that must already be in the file
-        /// for the patch to apply. The patch content is inserted on the
-        /// next line after the first occurrence. If absent and `after`
-        /// is set, the step fails (file structure changed). When `after`
-        /// is `None`, the patch is appended to the end of the file.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        after: Option<String>,
-        /// Patch content to insert. Must contain `marker` somewhere so
-        /// future runs detect it as already-applied — the step validates
-        /// this at apply time.
-        content: String,
-        /// Operation: `apply` (mutate file) or `verify` (read-only check).
-        #[serde(default = "default_patch_op")]
-        op: PatchOp,
-        /// Human-readable label shown during execution.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        label: Option<String>,
-    },
-
-    /// Pre-flight / health check. Non-fatal in `up` (warns), fatal in `check`.
-    Check {
-        /// Optional stable node ID for dependency-aware pipeline ordering.
-        #[serde(default, rename = "id", skip_serializing_if = "Option::is_none")]
-        step_id: Option<String>,
-        /// Step IDs that must run before this step.
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        depends_on: Vec<String>,
-        /// Named preflight groups this check belongs to. Workload commands can
-        /// run only the groups required by a rig-owned workload while full
-        /// `rig check` continues to evaluate every check-pipeline step.
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        groups: Vec<String>,
-        /// Human-readable label.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        label: Option<String>,
-        /// The actual check spec.
-        #[serde(flatten)]
-        spec: CheckSpec,
-    },
-}
-
-fn default_patch_op() -> PatchOp {
-    PatchOp::Apply
-}
-
-/// Git operation supported by a rig `git` step.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "kebab-case")]
-pub enum GitOp {
-    /// `git status --porcelain=v1`. Passes if exit 0.
-    Status,
-    /// `git pull [args...]`.
-    Pull,
-    /// `git push [args...]`. Use `args` for `--force-with-lease`,
-    /// `--follow-tags`, etc. Plain `--force` is intentionally NOT
-    /// blocked at the rig layer — rigs can be reproduced or reverted, so
-    /// the safety boundary lives at the CLI surface.
-    Push,
-    /// `git fetch [args...]`.
-    Fetch,
-    /// `git checkout [args...]`.
-    Checkout,
-    /// `git rev-parse --abbrev-ref HEAD` — returns current branch in logs.
-    CurrentBranch,
-    /// `git rebase [<onto>]`. Default with no `args` rebases onto
-    /// `@{upstream}`. Use `args` to specify the upstream ref or extra
-    /// rebase flags.
-    Rebase,
-    /// `git cherry-pick <refs...>`. `args` is the list of commit refs to
-    /// pick (SHAs, branches, ranges). PR-number expansion via `gh` is a
-    /// CLI-only convenience and not modelled at the rig step level —
-    /// resolve PR numbers to SHAs in the rig spec.
-    CherryPick,
-}
-
-/// Stack operation supported by a rig `stack` step.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "kebab-case")]
-pub enum StackOp {
-    /// Delegate to `homeboy stack sync <stack-id>`.
-    Sync,
-}
-
-/// Service operation in a pipeline step.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "kebab-case")]
-pub enum ServiceOp {
-    Start,
-    Stop,
-    Health,
-}
-
-/// Symlink operation in a pipeline step.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "kebab-case")]
-pub enum SymlinkOp {
-    Ensure,
-    Verify,
-}
-
-/// Shared path operation in a pipeline step.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "kebab-case")]
-pub enum SharedPathOp {
-    /// Create missing dependency paths as symlinks to their shared targets.
-    Ensure,
-    /// Check that each dependency path is available without mutating anything.
-    Verify,
-    /// Remove only symlinks this rig created and still owns.
-    Cleanup,
-}
-
-/// Patch operation in a pipeline step.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "kebab-case")]
-pub enum PatchOp {
-    /// Apply the patch if its marker is absent. No-op if already applied.
-    Apply,
-    /// Read-only: pass if the marker is present, fail otherwise. Use in
-    /// `check` pipelines to surface stale or unpatched checkouts.
-    Verify,
-}
-
-/// A single declarative check. One-of semantics — exactly one of the
-/// probe fields (`http`, `file`, `any_file_exists`, `command`, `newer_than`)
-/// should be set.
-/// Validated at check-time, not parse-time, because serde flattening
-/// across tagged enums is awkward and explicit-field checks keep the
-/// spec readable.
-#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct CheckSpec {
-    /// HTTP GET — passes if status matches `expect_status` (default 200).
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub http: Option<String>,
-
-    /// Expected HTTP status for the `http` check.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub expect_status: Option<u16>,
-
-    /// File path — passes if the file exists.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub file: Option<String>,
-
-    /// If set along with `file`, also requires the file contents to contain
-    /// this substring. Cheap probe for verifying drop-ins / generated files.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub contains: Option<String>,
-
-    /// File paths — passes when at least one path exists.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub any_file_exists: Vec<String>,
-
-    /// Shell command — passes if exit code matches `expect_exit` (default 0).
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub command: Option<String>,
-
-    /// Expected exit code for the `command` check.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub expect_exit: Option<i32>,
-
-    /// Mtime / staleness comparison — passes when `left` is newer than
-    /// `right`. Surfaces "I rebuilt but the daemon is still on the old
-    /// bundle" failures the wiki preflight calls out as the #1 dev-env
-    /// confusion source. If the `process_start` source resolves to no
-    /// running process, the check passes (no stale daemon to recycle).
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub newer_than: Option<NewerThanSpec>,
-}
-
-/// Mtime / staleness comparison check.
-///
-/// Each side picks one source. `left > right` ⇒ pass. Equal or `left < right`
-/// ⇒ fail. "Source missing" semantics differ by side: if `left` is a
-/// `process_start` and no process matches, the check passes (interpretation:
-/// no stale daemon to fight with). Any other missing source is an error.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct NewerThanSpec {
-    pub left: TimeSource,
-    pub right: TimeSource,
-}
-
-/// A time source for `newer_than` checks. One-of semantics enforced at
-/// evaluate-time.
-#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct TimeSource {
-    /// File mtime (seconds since epoch). Path supports `~` and `${...}`
-    /// expansion.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub file_mtime: Option<String>,
-
-    /// Process start time (seconds since epoch). Discovers the newest
-    /// matching process by command-line substring (`ps -o args`). When no
-    /// process matches and this source is on the `left`, the parent check
-    /// passes — there's no stale process to flag.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub process_start: Option<DiscoverSpec>,
 }
 
 #[cfg(test)]
